[modmail] move class to different package
